add standby timer processing logic, separate existing timer processing logic into active timer processing logic file
