+++ service/history/historyEngine.go
@@ -394,7 +394,7 @@ func (e *historyEngineImpl) StartWorkflowExecution(startRequest *h.StartWorkflow
 	}
 
 	if err == nil {
-		e.timerProcessor.NotifyNewTimers(timerTasks)
+		e.timerProcessor.NotifyNewTimers(e.shard.GetService().GetClusterMetadata().GetCurrentClusterName(), timerTasks)
 
 		return &workflow.StartWorkflowExecutionResponse{
 			RunId: common.StringPtr(resultRunID),
@@ -670,7 +670,7 @@ Update_History_Loop:
 		// Start a timer for the decision task.
 		timeOutTask := tBuilder.AddDecisionTimoutTask(scheduleID, di.Attempt, di.DecisionTimeout)
 		timerTasks := []persistence.Task{timeOutTask}
-		defer e.timerProcessor.NotifyNewTimers(timerTasks)
+		defer e.timerProcessor.NotifyNewTimers(e.shard.GetService().GetClusterMetadata().GetCurrentClusterName(), timerTasks)
 
 		// Generate a transaction ID for appending events to history
 		transactionID, err2 := e.shard.GetNextTransferTaskID()
@@ -1268,7 +1268,7 @@ Update_History_Loop:
 		// add continueAsNewTimerTask
 		timerTasks = append(timerTasks, continueAsNewTimerTasks...)
 		// Inform timer about the new ones.
-		e.timerProcessor.NotifyNewTimers(timerTasks)
+		e.timerProcessor.NotifyNewTimers(e.shard.GetService().GetClusterMetadata().GetCurrentClusterName(), timerTasks)
 
 		return err
 	}
@@ -1832,7 +1832,7 @@ Update_History_Loop:
 			}
 			return err
 		}
-		e.timerProcessor.NotifyNewTimers(timerTasks)
+		e.timerProcessor.NotifyNewTimers(e.shard.GetService().GetClusterMetadata().GetCurrentClusterName(), timerTasks)
 		return nil
 	}
 	return ErrMaxAttemptsExceeded
