+++ service/history/timerQueueActiveProcessor.go
@@ -1,0 +1,489 @@
+// Copyright (c) 2017 Uber Technologies, Inc.
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in
+// all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+// THE SOFTWARE.
+
+package history
+
+import (
+	"fmt"
+	"time"
+
+	"github.com/uber-common/bark"
+	workflow "github.com/uber/cadence/.gen/go/shared"
+	"github.com/uber/cadence/common/logging"
+	"github.com/uber/cadence/common/metrics"
+	"github.com/uber/cadence/common/persistence"
+)
+
+type (
+	timerQueueActiveProcessorImpl struct {
+		shard                   ShardContext
+		historyService          *historyEngineImpl
+		cache                   *historyCache
+		logger                  bark.Logger
+		metricsClient           metrics.Client
+		timerQueueProcessorBase *timerQueueProcessorBase
+		timerQueueAckMgr        timerQueueAckMgr
+	}
+)
+
+func newTimerQueueActiveProcessor(shard ShardContext, historyService *historyEngineImpl, executionManager persistence.ExecutionManager, logger bark.Logger) *timerQueueActiveProcessorImpl {
+	log := logger.WithFields(bark.Fields{
+		logging.TagWorkflowComponent: logging.TagValueTimerQueueComponent,
+	})
+	processor := &timerQueueActiveProcessorImpl{
+		shard:                   shard,
+		historyService:          historyService,
+		cache:                   historyService.historyCache,
+		logger:                  log,
+		metricsClient:           historyService.metricsClient,
+		timerQueueProcessorBase: newTimerQueueProcessorBase(shard, historyService, executionManager, logger),
+		timerQueueAckMgr:        newTimerQueueAckMgr(shard, historyService.metricsClient, executionManager, shard.GetService().GetClusterMetadata().GetCurrentClusterName(), log),
+	}
+	processor.timerQueueProcessorBase.timerProcessor = processor
+	return processor
+}
+
+func (t *timerQueueActiveProcessorImpl) process(timerTask *persistence.TimerTaskInfo) error {
+	taskID := TimerSequenceID{VisibilityTimestamp: timerTask.VisibilityTimestamp, TaskID: timerTask.TaskID}
+	t.logger.Debugf("Processing timer: (%s), for WorkflowID: %v, RunID: %v, Type: %v, TimeoutType: %v, EventID: %v",
+		taskID, timerTask.WorkflowID, timerTask.RunID, t.timerQueueProcessorBase.getTimerTaskType(timerTask.TaskType),
+		workflow.TimeoutType(timerTask.TimeoutType).String(), timerTask.EventID)
+
+	var err error
+	scope := metrics.TimerQueueProcessorScope
+	switch timerTask.TaskType {
+	case persistence.TaskTypeUserTimer:
+		scope = metrics.TimerTaskUserTimerScope
+		err = t.processExpiredUserTimer(timerTask)
+
+	case persistence.TaskTypeActivityTimeout:
+		scope = metrics.TimerTaskActivityTimeoutScope
+		err = t.processActivityTimeout(timerTask)
+
+	case persistence.TaskTypeDecisionTimeout:
+		scope = metrics.TimerTaskDecisionTimeoutScope
+		err = t.processDecisionTimeout(timerTask)
+
+	case persistence.TaskTypeWorkflowTimeout:
+		scope = metrics.TimerTaskWorkflowTimeoutScope
+		err = t.processWorkflowTimeout(timerTask)
+
+	case persistence.TaskTypeDeleteHistoryEvent:
+		scope = metrics.TimerTaskDeleteHistoryEvent
+		err = t.timerQueueProcessorBase.processDeleteHistoryEvent(timerTask)
+	}
+
+	if err != nil {
+		if _, ok := err.(*workflow.EntityNotExistsError); ok {
+			// Timer could fire after the execution is deleted.
+			// In which case just ignore the error so we can complete the timer task.
+			t.timerQueueAckMgr.completeTimerTask(timerTask)
+			err = nil
+		}
+		if err != nil {
+			t.metricsClient.IncCounter(scope, metrics.TaskFailures)
+		}
+	} else {
+		t.timerQueueAckMgr.completeTimerTask(timerTask)
+	}
+
+	return err
+}
+
+func (t *timerQueueActiveProcessorImpl) processExpiredUserTimer(task *persistence.TimerTaskInfo) error {
+	t.metricsClient.IncCounter(metrics.TimerTaskUserTimerScope, metrics.TaskRequests)
+	sw := t.metricsClient.StartTimer(metrics.TimerTaskUserTimerScope, metrics.TaskLatency)
+	defer sw.Stop()
+
+	context, release, err0 := t.cache.getOrCreateWorkflowExecution(t.timerQueueProcessorBase.getDomainIDAndWorkflowExecution(task))
+	if err0 != nil {
+		return err0
+	}
+	defer release()
+
+Update_History_Loop:
+	for attempt := 0; attempt < conditionalRetryCount; attempt++ {
+		msBuilder, err1 := context.loadWorkflowExecution()
+		if err1 != nil {
+			return err1
+		}
+		tBuilder := t.historyService.getTimerBuilder(&context.workflowExecution)
+
+		if !msBuilder.isWorkflowExecutionRunning() {
+			// Workflow is completed.
+			return nil
+		}
+
+		var timerTasks []persistence.Task
+		scheduleNewDecision := false
+
+	ExpireUserTimers:
+		for _, td := range tBuilder.GetUserTimers(msBuilder) {
+			hasTimer, ti := tBuilder.GetUserTimer(td.TimerID)
+			if !hasTimer {
+				t.logger.Debugf("Failed to find in memory user timer: %s", td.TimerID)
+				return fmt.Errorf("Failed to find in memory user timer: %s", td.TimerID)
+			}
+
+			if isExpired := tBuilder.IsTimerExpired(td, task.VisibilityTimestamp); isExpired {
+				// Add TimerFired event to history.
+				if msBuilder.AddTimerFiredEvent(ti.StartedID, ti.TimerID) == nil {
+					return errFailedToAddTimerFiredEvent
+				}
+
+				scheduleNewDecision = !msBuilder.HasPendingDecisionTask()
+			} else {
+				// See if we have next timer in list to be created.
+				if !td.TaskCreated {
+					nextTask := tBuilder.createNewTask(td)
+					timerTasks = []persistence.Task{nextTask}
+
+					// Update the task ID tracking the corresponding timer task.
+					ti.TaskID = nextTask.GetTaskID()
+					msBuilder.UpdateUserTimer(ti.TimerID, ti)
+					defer t.timerQueueProcessorBase.NotifyNewTimers(timerTasks)
+				}
+
+				// Done!
+				break ExpireUserTimers
+			}
+		}
+
+		// We apply the update to execution using optimistic concurrency.  If it fails due to a conflict than reload
+		// the history and try the operation again.
+		err := t.updateWorkflowExecution(context, msBuilder, scheduleNewDecision, false, timerTasks, nil)
+		if err != nil {
+			if err == ErrConflict {
+				continue Update_History_Loop
+			}
+		}
+		return err
+	}
+	return ErrMaxAttemptsExceeded
+}
+
+func (t *timerQueueActiveProcessorImpl) processActivityTimeout(timerTask *persistence.TimerTaskInfo) error {
+	t.metricsClient.IncCounter(metrics.TimerTaskActivityTimeoutScope, metrics.TaskRequests)
+	sw := t.metricsClient.StartTimer(metrics.TimerTaskActivityTimeoutScope, metrics.TaskLatency)
+	defer sw.Stop()
+
+	context, release, err0 := t.cache.getOrCreateWorkflowExecution(t.timerQueueProcessorBase.getDomainIDAndWorkflowExecution(timerTask))
+	if err0 != nil {
+		return err0
+	}
+	defer release()
+
+Update_History_Loop:
+	for attempt := 0; attempt < conditionalRetryCount; attempt++ {
+		msBuilder, err1 := context.loadWorkflowExecution()
+		if err1 != nil {
+			return err1
+		}
+		tBuilder := t.historyService.getTimerBuilder(&context.workflowExecution)
+
+		scheduleID := timerTask.EventID
+		// First check to see if cache needs to be refreshed as we could potentially have stale workflow execution in
+		// some extreme cassandra failure cases.
+		if scheduleID >= msBuilder.GetNextEventID() {
+			t.metricsClient.IncCounter(metrics.TimerQueueProcessorScope, metrics.StaleMutableStateCounter)
+			t.logger.Debugf("processActivityTimeout: scheduleID mismatch. MS NextEventID: %v, scheduleID: %v",
+				msBuilder.GetNextEventID(), scheduleID)
+			// Reload workflow execution history
+			context.clear()
+			continue Update_History_Loop
+		}
+
+		if !msBuilder.isWorkflowExecutionRunning() {
+			// Workflow is completed.
+			return nil
+		}
+
+		var timerTasks []persistence.Task
+		updateHistory := false
+		createNewTimer := false
+
+	ExpireActivityTimers:
+		for _, td := range tBuilder.GetActivityTimers(msBuilder) {
+			ai, isRunning := msBuilder.GetActivityInfo(td.ActivityID)
+			if !isRunning {
+				//  We might have time out this activity already.
+				continue ExpireActivityTimers
+			}
+
+			if isExpired := tBuilder.IsTimerExpired(td, timerTask.VisibilityTimestamp); isExpired {
+				timeoutType := td.TimeoutType
+				t.logger.Debugf("Activity TimeoutType: %v, scheduledID: %v, startedId: %v. \n",
+					timeoutType, ai.ScheduleID, ai.StartedID)
+
+				switch timeoutType {
+				case workflow.TimeoutTypeScheduleToClose:
+					{
+						t.metricsClient.IncCounter(metrics.TimerTaskActivityTimeoutScope, metrics.ScheduleToCloseTimeoutCounter)
+						if msBuilder.AddActivityTaskTimedOutEvent(ai.ScheduleID, ai.StartedID, timeoutType, nil) == nil {
+							return errFailedToAddTimeoutEvent
+						}
+						updateHistory = true
+					}
+
+				case workflow.TimeoutTypeStartToClose:
+					{
+						t.metricsClient.IncCounter(metrics.TimerTaskActivityTimeoutScope, metrics.StartToCloseTimeoutCounter)
+						if ai.StartedID != emptyEventID {
+							if msBuilder.AddActivityTaskTimedOutEvent(ai.ScheduleID, ai.StartedID, timeoutType, nil) == nil {
+								return errFailedToAddTimeoutEvent
+							}
+							updateHistory = true
+						}
+					}
+
+				case workflow.TimeoutTypeHeartbeat:
+					{
+						t.metricsClient.IncCounter(metrics.TimerTaskActivityTimeoutScope, metrics.HeartbeatTimeoutCounter)
+						t.logger.Debugf("Activity Heartbeat expired: %+v", *ai)
+
+						if msBuilder.AddActivityTaskTimedOutEvent(ai.ScheduleID, ai.StartedID, timeoutType, ai.Details) == nil {
+							return errFailedToAddTimeoutEvent
+						}
+						updateHistory = true
+					}
+
+				case workflow.TimeoutTypeScheduleToStart:
+					{
+						t.metricsClient.IncCounter(metrics.TimerTaskActivityTimeoutScope, metrics.ScheduleToStartTimeoutCounter)
+						if ai.StartedID == emptyEventID {
+							if msBuilder.AddActivityTaskTimedOutEvent(ai.ScheduleID, ai.StartedID, timeoutType, nil) == nil {
+								return errFailedToAddTimeoutEvent
+							}
+							updateHistory = true
+						}
+					}
+				}
+			} else {
+				// See if we have next timer in list to be created.
+				isHeartBeatTask := timerTask.TimeoutType == int(workflow.TimeoutTypeHeartbeat)
+
+				// Create next timer task if we don't have one (or)
+				// if current one is HB task and we need to create next HB task for the same.
+				// NOTE: When record activity HB comes in we only update last heartbeat timestamp, this is the place
+				// where we create next timer task based on that new updated timestamp.
+				if !td.TaskCreated || (isHeartBeatTask && td.EventID == scheduleID) {
+					nextTask := tBuilder.createNewTask(td)
+					timerTasks = []persistence.Task{nextTask}
+					at := nextTask.(*persistence.ActivityTimeoutTask)
+
+					ai.TimerTaskStatus = ai.TimerTaskStatus | getActivityTimerStatus(workflow.TimeoutType(at.TimeoutType))
+					msBuilder.UpdateActivity(ai)
+					createNewTimer = true
+
+					t.logger.Debugf("%s: Adding Activity Timeout: with timeout: %v sec, ExpiryTime: %s, TimeoutType: %v, EventID: %v",
+						time.Now(), td.TimeoutSec, at.VisibilityTimestamp, td.TimeoutType.String(), at.EventID)
+				}
+
+				// Done!
+				break ExpireActivityTimers
+			}
+		}
+
+		if updateHistory || createNewTimer {
+			// We apply the update to execution using optimistic concurrency.  If it fails due to a conflict than reload
+			// the history and try the operation again.
+			scheduleNewDecision := updateHistory && !msBuilder.HasPendingDecisionTask()
+			err := t.updateWorkflowExecution(context, msBuilder, scheduleNewDecision, false, timerTasks, nil)
+			if err != nil {
+				if err == ErrConflict {
+					continue Update_History_Loop
+				}
+			}
+
+			t.timerQueueProcessorBase.NotifyNewTimers(timerTasks)
+			return nil
+		}
+
+		return nil
+	}
+	return ErrMaxAttemptsExceeded
+}
+
+func (t *timerQueueActiveProcessorImpl) processDecisionTimeout(task *persistence.TimerTaskInfo) error {
+	t.metricsClient.IncCounter(metrics.TimerTaskDecisionTimeoutScope, metrics.TaskRequests)
+	sw := t.metricsClient.StartTimer(metrics.TimerTaskDecisionTimeoutScope, metrics.TaskLatency)
+	defer sw.Stop()
+
+	context, release, err0 := t.cache.getOrCreateWorkflowExecution(t.timerQueueProcessorBase.getDomainIDAndWorkflowExecution(task))
+	if err0 != nil {
+		return err0
+	}
+	defer release()
+
+Update_History_Loop:
+	for attempt := 0; attempt < conditionalRetryCount; attempt++ {
+		msBuilder, err1 := context.loadWorkflowExecution()
+		if err1 != nil {
+			return err1
+		}
+
+		scheduleID := task.EventID
+		di, isPending := msBuilder.GetPendingDecision(scheduleID)
+
+		// First check to see if cache needs to be refreshed as we could potentially have stale workflow execution in
+		// some extreme cassandra failure cases.
+		if !isPending && scheduleID >= msBuilder.GetNextEventID() {
+			t.metricsClient.IncCounter(metrics.TimerQueueProcessorScope, metrics.StaleMutableStateCounter)
+			// Reload workflow execution history
+			context.clear()
+			continue Update_History_Loop
+		}
+
+		scheduleNewDecision := false
+		switch task.TimeoutType {
+		case int(workflow.TimeoutTypeStartToClose):
+			t.metricsClient.IncCounter(metrics.TimerTaskDecisionTimeoutScope, metrics.StartToCloseTimeoutCounter)
+			if isPending && di.Attempt == task.ScheduleAttempt && msBuilder.isWorkflowExecutionRunning() {
+				// Add a decision task timeout event.
+				msBuilder.AddDecisionTaskTimedOutEvent(scheduleID, di.StartedID)
+				scheduleNewDecision = true
+			}
+		case int(workflow.TimeoutTypeScheduleToStart):
+			t.metricsClient.IncCounter(metrics.TimerTaskDecisionTimeoutScope, metrics.ScheduleToStartTimeoutCounter)
+			// decision schedule to start timeout only apply to sticky decision
+			// check if scheduled decision still pending and not started yet
+			if isPending && di.Attempt == task.ScheduleAttempt && msBuilder.isWorkflowExecutionRunning() &&
+				di.StartedID == emptyEventID && msBuilder.isStickyTaskListEnabled() {
+				timeoutEvent := msBuilder.AddDecisionTaskScheduleToStartTimeoutEvent(scheduleID)
+				if timeoutEvent == nil {
+					// Unable to add DecisionTaskTimedout event to history
+					return &workflow.InternalServiceError{Message: "Unable to add DecisionTaskScheduleToStartTimeout event to history."}
+				}
+
+				// reschedule decision, which will be on its original task list
+				scheduleNewDecision = true
+			}
+		}
+
+		if scheduleNewDecision {
+			// We apply the update to execution using optimistic concurrency.  If it fails due to a conflict than reload
+			// the history and try the operation again.
+			err := t.updateWorkflowExecution(context, msBuilder, scheduleNewDecision, false, nil, nil)
+			if err != nil {
+				if err == ErrConflict {
+					continue Update_History_Loop
+				}
+			}
+			return err
+		}
+
+		return nil
+
+	}
+	return ErrMaxAttemptsExceeded
+}
+
+func (t *timerQueueActiveProcessorImpl) processWorkflowTimeout(task *persistence.TimerTaskInfo) error {
+	t.metricsClient.IncCounter(metrics.TimerTaskWorkflowTimeoutScope, metrics.TaskRequests)
+	sw := t.metricsClient.StartTimer(metrics.TimerTaskWorkflowTimeoutScope, metrics.TaskLatency)
+	defer sw.Stop()
+
+	context, release, err0 := t.cache.getOrCreateWorkflowExecution(t.timerQueueProcessorBase.getDomainIDAndWorkflowExecution(task))
+	if err0 != nil {
+		return err0
+	}
+	defer release()
+
+Update_History_Loop:
+	for attempt := 0; attempt < conditionalRetryCount; attempt++ {
+		msBuilder, err1 := context.loadWorkflowExecution()
+		if err1 != nil {
+			return err1
+		}
+
+		if !msBuilder.isWorkflowExecutionRunning() {
+			return nil
+		}
+
+		if e := msBuilder.AddTimeoutWorkflowEvent(); e == nil {
+			// If we failed to add the event that means the workflow is already completed.
+			// we drop this timeout event.
+			return nil
+		}
+
+		// We apply the update to execution using optimistic concurrency.  If it fails due to a conflict than reload
+		// the history and try the operation again.
+		err := t.updateWorkflowExecution(context, msBuilder, false, true, nil, nil)
+		if err != nil {
+			if err == ErrConflict {
+				continue Update_History_Loop
+			}
+		}
+		return err
+	}
+	return ErrMaxAttemptsExceeded
+}
+
+func (t *timerQueueActiveProcessorImpl) updateWorkflowExecution(
+	context *workflowExecutionContext,
+	msBuilder *mutableStateBuilder,
+	scheduleNewDecision bool,
+	createDeletionTask bool,
+	timerTasks []persistence.Task,
+	clearTimerTask persistence.Task,
+) error {
+	var transferTasks []persistence.Task
+	if scheduleNewDecision {
+		// Schedule a new decision.
+		di := msBuilder.AddDecisionTaskScheduledEvent()
+		transferTasks = []persistence.Task{&persistence.DecisionTask{
+			DomainID:   msBuilder.executionInfo.DomainID,
+			TaskList:   di.Tasklist,
+			ScheduleID: di.ScheduleID,
+		}}
+		if msBuilder.isStickyTaskListEnabled() {
+			tBuilder := t.historyService.getTimerBuilder(&context.workflowExecution)
+			stickyTaskTimeoutTimer := tBuilder.AddScheduleToStartDecisionTimoutTask(di.ScheduleID, di.Attempt,
+				msBuilder.executionInfo.StickyScheduleToStartTimeout)
+			timerTasks = append(timerTasks, stickyTaskTimeoutTimer)
+		}
+	}
+
+	if createDeletionTask {
+		tBuilder := t.historyService.getTimerBuilder(&context.workflowExecution)
+		tranT, timerT, err := t.historyService.getDeleteWorkflowTasks(msBuilder.executionInfo.DomainID, tBuilder)
+		if err != nil {
+			return nil
+		}
+		transferTasks = append(transferTasks, tranT)
+		timerTasks = append(timerTasks, timerT)
+	}
+
+	// Generate a transaction ID for appending events to history
+	transactionID, err1 := t.historyService.shard.GetNextTransferTaskID()
+	if err1 != nil {
+		return err1
+	}
+
+	err := context.updateWorkflowExecutionWithDeleteTask(transferTasks, timerTasks, clearTimerTask, transactionID)
+	if err != nil {
+		if isShardOwnershiptLostError(err) {
+			// Shard is stolen.  Stop timer processing to reduce duplicates
+			t.timerQueueProcessorBase.Stop()
+		}
+	}
+	t.timerQueueProcessorBase.NotifyNewTimers(timerTasks)
+	return err
+}
